+++ src/ninja/mbedded/ninjaterm/model/status/Status.java
@@ -1,11 +1,17 @@
 package ninja.mbedded.ninjaterm.model.status;
 
-import javafx.beans.property.ListProperty;
+import javafx.animation.Animation;
+import javafx.animation.KeyFrame;
+import javafx.animation.Timeline;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.Node;
 import javafx.scene.paint.Color;
 import javafx.scene.text.Text;
+import javafx.util.Duration;
+import ninja.mbedded.ninjaterm.model.Model;
+import ninja.mbedded.ninjaterm.model.terminal.Terminal;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -18,14 +24,47 @@ import java.util.Date;
  *
  * @author          Geoffrey Hunter <gbmhunter@gmail.com> (www.mbedded.ninja)
  * @since           2016-09-18
- * @last-modified   2016-09-23
+ * @last-modified   2016-10-07
  */
 public class Status {
 
+    private static final double BITS_PER_SECOND_CALC_PERIOD_MS = 1000.0;
+
+    private Model model;
+
     public ObservableList<Node> statusMsgs = FXCollections.observableList(new ArrayList<Node>());
 
-    public Status() {
+    public SimpleDoubleProperty totalBytesPerSecTx = new SimpleDoubleProperty();
+    public SimpleDoubleProperty totalBytesPerSecRx = new SimpleDoubleProperty();
 
+    public Status(Model model) {
+
+        this.model = model;
+
+        // Setup timer to trigger calculation of bits/second at a fixed rate
+        Timeline timeline = new Timeline(new KeyFrame(
+                Duration.millis(BITS_PER_SECOND_CALC_PERIOD_MS),
+                ae -> calculateBytesPerSec()));
+        timeline.setCycleCount(Animation.INDEFINITE);
+        timeline.play();
+
+    }
+
+    /**
+     * This should be called once every <code>BITS_PER_SECOND_CALC_PERIOD_MS</code>.
+     */
+    private void calculateBytesPerSec() {
+
+        double bytesPerSecondSumTx = 0.0;
+        double bytesPerSecondSumRx = 0.0;
+
+        for(Terminal terminal : model.terminals) {
+            bytesPerSecondSumTx += terminal.stats.bytesPerSecondTx.get();
+            bytesPerSecondSumRx += terminal.stats.bytesPerSecondRx.get();
+        }
+
+        totalBytesPerSecTx.set(bytesPerSecondSumTx);
+        totalBytesPerSecRx.set(bytesPerSecondSumRx);
     }
 
     /**