+++ src/ninja/mbedded/ninjaterm/model/terminal/stats/Stats.java
@@ -17,46 +17,46 @@ import ninja.mbedded.ninjaterm.model.terminal.Terminal;
  */
 public class Stats {
 
-    private static final double BITS_PER_SECOND_CALC_PERIOD_MS = 1000.0;
+    private static final double BYTES_PER_SECOND_CALC_PERIOD_MS = 1000.0;
 
     public SimpleIntegerProperty numCharactersTx = new SimpleIntegerProperty(0);
     public SimpleIntegerProperty numCharactersRx = new SimpleIntegerProperty(0);
 
-    private double bitsSinceLastCalcRx = 0.0;
-    private double bitsSinceLastCalcTx = 0.0;
+    private double bytesSinceLastCalcRx = 0.0;
+    private double bytesSinceLastCalcTx = 0.0;
 
-    public SimpleDoubleProperty bitsPerSecondTx = new SimpleDoubleProperty(0.0);
-    public SimpleDoubleProperty bitsPerSecondRx = new SimpleDoubleProperty(0.0);
+    public SimpleDoubleProperty bytesPerSecondTx = new SimpleDoubleProperty(0.0);
+    public SimpleDoubleProperty bytesPerSecondRx = new SimpleDoubleProperty(0.0);
 
     public Stats(Terminal terminal) {
 
         // INSTALL TX/RX DATA LISTENERS
         terminal.txRx.dataSentTxListeners.add(txData -> {
-            bitsSinceLastCalcTx += txData.length();
+            bytesSinceLastCalcTx += txData.length();
         });
 
         terminal.txRx.rawDataReceivedListeners.add(data -> {
-            bitsSinceLastCalcRx += data.length();
+            bytesSinceLastCalcRx += data.length();
         });
 
         // Setup timer to trigger calculation of bits/second at a fixed rate
         Timeline timeline = new Timeline(new KeyFrame(
-                Duration.millis(BITS_PER_SECOND_CALC_PERIOD_MS),
-                ae -> recalculateBitsPerSecond()));
+                Duration.millis(BYTES_PER_SECOND_CALC_PERIOD_MS),
+                ae -> calculateBytesPerSecond()));
         timeline.setCycleCount(Animation.INDEFINITE);
         timeline.play();
     }
 
     /**
-     * This should be called once every <code>BITS_PER_SECOND_CALC_PERIOD_MS</code>.
+     * This should be called once every <code>BYTES_PER_SECOND_CALC_PERIOD_MS</code>.
      */
-    private void recalculateBitsPerSecond() {
+    private void calculateBytesPerSecond() {
 
-        bitsPerSecondTx.set(bitsSinceLastCalcTx/ (BITS_PER_SECOND_CALC_PERIOD_MS/1000.0));
-        bitsSinceLastCalcTx = 0.0;
+        bytesPerSecondTx.set(bytesSinceLastCalcTx / (BYTES_PER_SECOND_CALC_PERIOD_MS /1000.0));
+        bytesSinceLastCalcTx = 0.0;
 
-        bitsPerSecondRx.set(bitsSinceLastCalcRx/ (BITS_PER_SECOND_CALC_PERIOD_MS/1000.0));
-        bitsSinceLastCalcRx = 0.0;
+        bytesPerSecondRx.set(bytesSinceLastCalcRx / (BYTES_PER_SECOND_CALC_PERIOD_MS /1000.0));
+        bytesSinceLastCalcRx = 0.0;
     }
 
 }