+++ changelog.md
@@ -6,6 +6,7 @@ vX.X.X
 - "Parse ANSI escape sequences" checkbox is now only enabled when decoding mode is ASCII, closes #117.
 - Fixed bug where ANSI escape codes were not disabled when it's checkbox is deselected, closes #119.
 - Added bits/second metrics to the "Stats" sub-tab, closes #120.
+- Added bottom bar to the status pane which displays the total TX/RX bytes per second rates, closes #122.
 
 v0.6.2
 ------